Create README.md
Abstract
he project entails using a classification algorithm to build a model from the historical data of patient and their response to varied medications. The project will us the trained decision tree to predict the class of the unknown patient or prescribe the proper drug for the respective patient. 

Introduction
The dataset contains the descriptions of many patients who are suffering from the same illness. In the course of undertaking their medications, the patients are monitored to acertain if they respond well to the drugs. For instance the patient will be given drugA,drugB, drugC and drugD to establish if they respond to the same illness. Part of this assignment is to investigate whether the prescribed drugs is effective and efficiency to be administered to the patient suffering from same illness in future. This is sample example of multiclass classifier that entails training part of the the decision tree and use it to predict the class of the unknown patient or prescribe the appropriate drug for the patient. The assignment will entail;
1. Importing of necessary libraries. 
2. Data preprocessing
3. Setting up of the Decision tree classifier
4. Modelling
5. Prediction
6. Eavaluation
7. Visualization